+++ gulpfile.js
@@ -1,5 +1,3 @@
-'use strict';
-
 /*
   gulpfile.js
   ===========
@@ -12,7 +10,26 @@
   when you run `gulp`.
 */
 
-var requireDir = require( 'require-dir' );
+const gulp = require( 'gulp' );
+const requireDir = require( 'require-dir' );
 
-// Require all tasks in gulp/tasks, including subfolders
+// Require all tasks in gulp/tasks, including subfolders.
 requireDir( './gulp/tasks', { recurse: true } );
+
+gulp.task( 'build',
+  gulp.series(
+    gulp.parallel(
+      'styles',
+      'scripts',
+      'images'
+    ),
+    'copy'
+  )
+);
+
+gulp.task( 'default',
+  gulp.parallel(
+    'lint',
+    'build'
+  )
+);