+++ src/disclosures/js/dispatchers/post-verify.js
@@ -1,25 +1,26 @@
-'use strict';
+// TODO: Remove jquery.
+const $ = require( 'jquery' );
 
-var postVerify = {
+const postVerify = {
   csrfToken: null,
 
   init: function() {
-    // Alternate code. getting token from document.cookie:
-    // if ( document.cookie && document.cookie != '' ) {
-    //   var cookies = document.cookie.split( ';' );
-    //   for ( var x = 0; x < cookes.length; x++ ) {
-    //     var cookie = $.trim( cookies[ x ] );
-    //     if ( cookie.substring( 0, 10 ) === 'csrftoken=' ) {
-    //       this.csrfToken = cookie.substring( 10 );
-    //       break;
-    //     }
-    //   }
-    // }
+    /* Alternate code. getting token from document.cookie:
+       if ( document.cookie && document.cookie != '' ) {
+       var cookies = document.cookie.split( ';' );
+       for ( var x = 0; x < cookes.length; x++ ) {
+       var cookie = $.trim( cookies[ x ] );
+       if ( cookie.substring( 0, 10 ) === 'csrftoken=' ) {
+       this.csrfToken = cookie.substring( 10 );
+       break;
+       }
+       }
+       } */
     this.csrfToken = $( '[name="csrfmiddlewaretoken"]' ).val();
   },
 
   verify: function( offerID, collegeID, error ) {
-    var url,
+    let url,
         postdata = {
           csrfmiddlewaretoken:  this.csrfToken,
           oid:                  offerID,