+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## Unreleased
+
+### Added
+-
+
+### Changed
+- Convert browserify config to webpack.
+- Remove unused dependencies.
+- Update ESLint config to version 5.
+- Fix linter errors.
+- Convert to ES6 syntax.
+
+### Removed
+-
+
 ## 2.3.6
 - Modified Google Analytics question event to fire independent of settlement status